remove fam from set, transplant fam to set, combat mode toggle on salute and .combat, fixed crafting durability bonus (integer division strikes again x_x) and fixed fam stats to be more like what I intended (damage to units starts at 10% of their base power and rises to 100% at max)

+++ Patches/EquipmentPatches.cs
@@ -391,7 +391,7 @@ internal static class EquipmentPatches
                                     {
                                         Durability durability = itemEntity.Read<Durability>();
                                         int level = handler.GetExperienceData(steamId).Key;
-                                        durability.MaxDurability *= (1 + level / Plugin.MaxProfessionLevel.Value);
+                                        durability.MaxDurability *= (1 + (float)level / (float)Plugin.MaxProfessionLevel.Value);
                                         durability.Value = durability.MaxDurability;
                                         itemEntity.Write(durability);
                                     }
@@ -403,7 +403,7 @@ internal static class EquipmentPatches
                                     {
                                         Durability durability = itemEntity.Read<Durability>();
                                         int level = handler.GetExperienceData(steamId).Key;
-                                        durability.MaxDurability *= (1 + level / Plugin.MaxProfessionLevel.Value);
+                                        durability.MaxDurability *= (1 + (float)level / (float)Plugin.MaxProfessionLevel.Value);
                                         durability.Value = durability.MaxDurability;
                                         itemEntity.Write(durability);
                                     }
@@ -413,7 +413,7 @@ internal static class EquipmentPatches
                                     {
                                         Durability durability = itemEntity.Read<Durability>();
                                         int level = handler.GetExperienceData(steamId).Key;
-                                        durability.MaxDurability *= (1 + level / Plugin.MaxProfessionLevel.Value);
+                                        durability.MaxDurability *= (1 + (float)level / (float)Plugin.MaxProfessionLevel.Value);
                                         durability.Value = durability.MaxDurability;
                                         itemEntity.Write(durability);
                                     }
